[#25] ProductService 위치 변경
[작업 내용]
- ProductService.java의 경로를 supermarket-api-product/src/main/java/com/harmony/supermarketapiproduct/application으로 변경

[테스트 내용]
- ProductService.java의 경로 변경 이후 스프링 부트 애플리케이션 실행 및 정상 동작 확인
